\name{CorTree-package}
\alias{CorTree-package}
\alias{CorTree}
\docType{package}
\title{\packageTitle{CorTree}}
\description{\packageDescription{CorTree}}
\section{Package Content}{\packageIndices{CorTree}}
\author{\packageAuthor{CorTree}}
\section{Maintainer}{\packageMaintainer{CorTree}}
\keyword{package}
